apps/slack/src/inngest/functions/slack/event-handlers/channel-unarchive.test.ts
import { expect, test, describe, beforeAll, afterAll, vi, afterEach } from 'vitest';
import * as slack from 'slack-web-api-client';
import type { SlackEvent } from '@slack/bolt';
import { createFunctionMock } from '@/inngest/__mocks__/inngest';
import { db } from '@/database/client';
import { teams } from '@/database/schema';
import { handleSlackWebhookEvent } from '../handle-slack-webhook-event';

vi.mock('slack-web-api-client');

const setup = createFunctionMock(handleSlackWebhookEvent, 'slack/webhook.handle');

const mockedDate = '2023-01-01T00:00:00.000Z';

const eventType: SlackEvent['type'] = 'channel_unarchive';

describe(`handle-slack-webhook-event ${eventType}`, () => {
  beforeAll(() => {
    vi.setSystemTime(mockedDate);
  });

  afterAll(() => {
    vi.useRealTimers();
  });

  afterEach(() => {
    vi.restoreAllMocks();
  });

  test('should insert channel successfully', async () => {
    const conversationInfoMock = vi.fn().mockResolvedValue({
      ok: true,
      channel: {
        id: 'channel-id-1',
        name: 'channel',
        is_ext_shared: false,
      },
      headers: new Headers(),
    } satisfies Awaited<ReturnType<typeof slack.SlackAPIClient.prototype.conversations.info>>);

    vi.spyOn(slack, 'SlackAPIClient').mockReturnValue({
      // @ts-expect-error -- this is a partial mock
      conversations: {
        info: conversationInfoMock,
      },
    });

    await db.insert(teams).values([
      {
        id: 'team-id',
        elbaOrganisationId: '00000000-0000-0000-0000-000000000001',
        token: 'token',
        url: 'https://url',
      },
    ]);

    const [result, { step }] = setup({
      team_id: 'team-id',
      // @ts-expect-error -- this is a partial mock
      event: {
        type: eventType,
        channel: 'channel-id-1',
      },
    });

    await expect(result).resolves.toStrictEqual({
      channelId: 'channel-id-1',
      message: 'Channel unarchived',
      teamId: 'team-id',
    });

    expect(slack.SlackAPIClient).toBeCalledTimes(1);
    expect(slack.SlackAPIClient).toBeCalledWith('token');

    expect(conversationInfoMock).toBeCalledTimes(1);
    expect(conversationInfoMock).toBeCalledWith({
      channel: 'channel-id-1',
    });

    const conversationsInserted = await db.query.conversations.findMany();

    expect(conversationsInserted).toEqual([
      {
        id: 'channel-id-1',
        isSharedExternally: false,
        lastSyncedAt: new Date(mockedDate),
        name: 'channel',
        teamId: 'team-id',
      },
    ]);

    expect(step.sendEvent).toBeCalledTimes(1);
    expect(step.sendEvent).toBeCalledWith('synchronize-conversation-messages', {
      data: {
        conversationId: 'channel-id-1',
        isFirstSync: false,
        teamId: 'team-id',
      },
      name: 'conversations/synchronize.messages',
    });
  });
});
